+++ locales/fi.js
@@ -1,0 +1,51 @@
+SirTrevor.Locales.fi = {
+  general: {
+    'delete':           'Poista?',
+    'drop':             'Raahaa __block__ tähän',
+    'paste':            'Tai lisää osoite tähän',
+    'upload':           '...tai valitse tiedosto',
+    'close':            'sulje',
+    'position':         'Sijainti',
+    'wait':             'Odota hetki...',
+    'link':             'Kirjoita osoite'
+  },
+  errors: {
+    'title': "Seuraavat virheet:",
+    'validation_fail': "__type__-lohko on epäkelpo",
+    'block_empty': "__name__ ei saa olla tyhjä",
+    'type_missing': "__type__ on pakollinen lohko",
+    'required_type_empty': "Pakollinen __type__-lohko on tyhjä",
+    'load_fail': "Sisällön lataamisessa on ongelma"
+  },
+  blocks: {
+    text: {
+      'title': "Teksti"
+    },
+    list: {
+      'title': "Lista"
+    },
+    quote: {
+      'title': "Lainaus",
+      'credit_field': "Lähde"
+    },
+    image: {
+      'title': "Kuva",
+      'upload_error': "Tiedoston lataamisessa oli ongelma"
+    },
+    video: {
+      'title': "Video"
+    },
+    tweet: {
+      'title': "Tweet",
+      'fetch_error': "Viestin hakemisessa oli ongelma"
+    },
+    embedly: {
+      'title': "Embedly",
+      'fetch_error': "Sisällön hakemisessa oli ongelma",
+      'key_missing': "Embedly vaatii API-avaimen"
+    },
+    heading: {
+      'title': 'Otsikko'
+    }
+  }
+};