+++ spec/javascripts/units/to_markdown.spec.js
@@ -168,4 +168,22 @@ describe("toMarkdown", function(){
     expect(markdown).toBe("[test](test)");
   });
 
+  it("strips all tags if aggresiveHTMLStrip is true", function() {
+    SirTrevor.DEFAULTS.toMarkdown.aggresiveHTMLStrip = true;
+
+    var html = "x < 4. y > 5.",
+        markdown = SirTrevor.toMarkdown(html, "Text");
+
+    expect(markdown).toBe("x  5.");
+  });
+
+  it("leaves < & > alone if aggresiveHTMLStrip is false", function(){
+    SirTrevor.DEFAULTS.toMarkdown.aggresiveHTMLStrip = false;
+
+    var html = "x < 4. y > 5.",
+        markdown = SirTrevor.toMarkdown(html, "Text");
+
+    expect(markdown).toBe("x < 4. y > 5.");
+  });
+
 });