+++ src/block_mixins/ajaxable.js
@@ -8,13 +8,17 @@ SirTrevor.BlockMixins.Ajaxable = {
     this._queued = [];
   },
 
-  addQueuedItem: function(name, deffered) {
+  addQueuedItem: function(name, deferred) {
     SirTrevor.log("Adding queued item for " + this.blockID + " called " + name);
-    this._queued.push({ name: name, deffered: deffered });
+    SirTrevor.EventBus.trigger("onUploadStart", this.blockID);
+
+    this._queued.push({ name: name, deferred: deferred });
   },
 
   removeQueuedItem: function(name) {
     SirTrevor.log("Removing queued item for " + this.blockID + " called " + name);
+    SirTrevor.EventBus.trigger("onUploadStop", this.blockID);
+
     this._queued = _.reject(this._queued, function(queued){ return queued.name == name; });
   },
 
@@ -25,7 +29,7 @@ SirTrevor.BlockMixins.Ajaxable = {
   resolveAllInQueue: function() {
     _.each(this._queued, function(item){
       SirTrevor.log("Aborting queued request: " + item.name);
-      item.deffered.abort();
+      item.deferred.abort();
     }, this);
   }