+++ src/extensions/file-uploader.js
@@ -5,8 +5,6 @@
 
 SirTrevor.fileUploader = function(block, file, success, error) {
 
-  SirTrevor.EventBus.trigger("onUploadStart");
-
   var uid  = [block.blockID, (new Date()).getTime(), 'raw'].join('-');
   var data = new FormData();
 
@@ -16,21 +14,19 @@ SirTrevor.fileUploader = function(block, file, success, error) {
 
   block.resetMessages();
 
-  var callbackSuccess = function(data){
+  var callbackSuccess = function(){
     SirTrevor.log('Upload callback called');
-    SirTrevor.EventBus.trigger("onUploadStop");
 
     if (!_.isUndefined(success) && _.isFunction(success)) {
-      _.bind(success, block)(data);
+      success.apply(block, arguments);
     }
   };
 
-  var callbackError = function(jqXHR, status, errorThrown){
+  var callbackError = function(){
     SirTrevor.log('Upload callback error called');
-    SirTrevor.EventBus.trigger("onUploadStop");
 
     if (!_.isUndefined(error) && _.isFunction(error)) {
-      _.bind(error, block)(status);
+      error.apply(block, arguments);
     }
   };