+++ src/simple-block.js
@@ -43,7 +43,7 @@ SirTrevor.SimpleBlock = (function(){
 
     type: '',
 
-    class: function() {
+    'class': function() {
       return _.classify(this.type);
     },
 
@@ -104,7 +104,7 @@ SirTrevor.SimpleBlock = (function(){
     },
 
     addMessage: function(msg, additionalClass) {
-      var $msg = $("<span>", { html: msg, class: "st-msg " + additionalClass });
+      var $msg = $("<span>", { html: msg, 'class': "st-msg " + additionalClass });
       this.$messages.append($msg)
                     .addClass('st-block__messages--is-visible');
       return $msg;